pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rivaflow"
version = "0.5.0"
description = "Training OS for the mat — log, analyze, improve"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Ruby Wolff" }
]
keywords = ["bjj", "jiu-jitsu", "training", "fitness", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Other/Nonlisted Topic",
]
# All dependencies pinned — aligned with rivaflow/requirements.txt
dependencies = [
    "fastapi==0.128.4",
    "uvicorn[standard]==0.40.0",
    "gunicorn==23.0.0",
    "typer==0.21.1",
    "rich==13.9.4",
    "pydantic==2.12.5",
    "pydantic-settings==2.12.0",
    "PyJWT[crypto]==2.10.1",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.3.0",
    "cryptography==46.0.5",
    "python-multipart==0.0.22",
    "httpx==0.28.1",
    "psycopg2==2.9.11",
    "pgvector==0.4.2",
    "python-dotenv==1.2.1",
    "email-validator==2.3.0",
    "slowapi==0.1.9",
    "redis==5.3.1",
    "sendgrid==6.12.5",
    "filetype==1.2.0",
    "apscheduler>=3.10,<4",
    "sentry-sdk[fastapi]==2.52.0",
    "boto3==1.36.14",
    "python-json-logger==3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-cov==6.0.0",
    "pytest-asyncio==0.25.2",
    "ruff==0.9.1",
    "pip-audit==2.7.3",
]

[project.scripts]
rivaflow = "rivaflow.cli.app:app"

[project.urls]
Homepage = "https://github.com/RubyWolff27/rivaflow"
Repository = "https://github.com/RubyWolff27/rivaflow"

[tool.hatch.build.targets.wheel]
packages = ["rivaflow"]

# NOTE: Root pyproject.toml uses line-length=88 (used by CI `black --check`).
# The inner rivaflow/pyproject.toml uses line-length=100 for local dev.
# Always run `black rivaflow/ tests/` from repo root to match CI.
[tool.ruff]
line-length = 88
target-version = "py311"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
exclude = ["build/"]
# Disable error codes for pre-existing issues; tighten incrementally
disable_error_code = [
    "arg-type",
    "assignment",
    "var-annotated",
    "union-attr",
    "call-arg",
    "index",
    "list-item",
    "attr-defined",
    "operator",
    "dict-item",
    "misc",
    "no-redef",
    "call-overload",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=rivaflow --cov-fail-under=55 --strict-markers"
pythonpath = ["rivaflow"]
